# Implement an optimized Cache::Entry coder