--- a/activesupport/lib/active_support/cache/memory_store.rb
+++ b/activesupport/lib/active_support/cache/memory_store.rb
@@ -45,8 +45,6 @@ module ActiveSupport
         end
       end
 
-      DEFAULT_CODER = DupCoder
-
       def initialize(options = nil)
         options ||= {}
         # Disable compression by default.
@@ -144,6 +142,10 @@ module ActiveSupport
 
       private
         PER_ENTRY_OVERHEAD = 240
+
+        def default_coder
+          DupCoder
+        end
 
         def cached_size(key, payload)
           key.to_s.bytesize + payload.bytesize + PER_ENTRY_OVERHEAD