--- a/activesupport/test/cache/stores/file_store_test.rb
+++ b/activesupport/test/cache/stores/file_store_test.rb
@@ -144,3 +144,36 @@ class FileStoreTest < ActiveSupport::TestCase
     assert_equal false, @cache.write(1, "aaaaaaaaaa", unless_exist: true)
   end
 end
+
+class OptimizedFileStoreTest < FileStoreTest
+  def setup
+    @previous_format = ActiveSupport::Cache.format_version
+    ActiveSupport::Cache.format_version = 7.0
+    super
+  end
+
+  def forward_compatibility
+    previous_format = ActiveSupport::Cache.format_version
+    ActiveSupport::Cache.format_version = 6.1
+    @old_store = lookup_store
+    ActiveSupport::Cache.format_version = previous_format
+
+    @old_store.write("foo", "bar")
+    assert_equal "bar", @cache.read("foo")
+  end
+
+  def forward_compatibility
+    previous_format = ActiveSupport::Cache.format_version
+    ActiveSupport::Cache.format_version = 6.1
+    @old_store = lookup_store
+    ActiveSupport::Cache.format_version = previous_format
+
+    @cache.write("foo", "bar")
+    assert_equal "bar", @old_store.read("foo")
+  end
+
+  def teardown
+    super
+    ActiveSupport::Cache.format_version = @previous_format
+  end
+end