@casperisfine 2021-04-20
NB: The major change since #40412, is that except for the "compression byte", all the header is marshaled and compressed. This is mostly for simplicity's sake. If anything, splitting the metadata (`expires_at`, `version`) would allow to rebuild the `Entry` object without decompressing the payload, which in some rare scenarios could improve performance.

But the PR is already much bigger than I'd like.