@casperisfine 2021-04-20
@fxn and @jeremy, since you were involved in https://github.com/rails/rails/issues/9494 (8 years ago...), maybe you'd be interested in looking at this PR?