@fxn 2021-04-21
I've done a superficial read of the patch, but need to do a better one.

This would need documentation too. What is the difference between the versions? If you have a cache full of payloads generated with version _X_, how do you migrate the project to version _X_ + 1? I believe this is cache docs, not upgrading docs, though upgrading docs for 7 could point to the cache docs for reference.

I'll write back!