@casperisfine 2021-04-21
> This would need documentation too.

Any pointer on similar docs I could take inspiration from?