@fxn 2021-04-21
> Any pointer on similar docs I could take inspiration from?

I think the natural place would be a new section in the [caching guide](https://guides.rubyonrails.org/caching_with_rails.html). Maybe it could be interesting to comment as well that the format version is not embedded in the payload, the reader could wonder, since Marshal payloads do. That also tell you that a transparent upgrade is not possible.

I do not know if you considered and discarded that possibility, would be just one additional byte with the version if we make it so. In that scenario the writer uses the configured format, and the reader is able select the loader depending on the embedded format version. Just thinking aloud at 23:39 :).