@casperisfine 2021-04-22
> I do not know if you considered and discarded that possibility, would be just one additional byte with the version if we make it so. In that scenario the writer uses the configured format, and the reader is able select the loader depending on the embedded format version.

That's actually what this PR does :)