@fxn 2021-04-22
> That's actually what this PR does :)

Where's the coder version encoded in the payload?