@casperisfine 2021-04-22
> Where's the coder version encoded in the payload?

So it's not the coder version itself, but:

  - Legacy payloads have the `Marshal` prefix, so `\x04....`
  - Newer uncompressed payloads have `\x00`
  - Newer compressed payloads have `\x01`

Both the legacy and the new encoder are capable for delegating to each other if they notice the prefix don't match.

So the upgrade is smooth as long as you switch encoder after upgrading to 7.0 (because 6.1 process don't know how to handle the new payloads).