@casperisfine 2021-04-22
There's even tests to make sure the transition is smooth: https://github.com/rails/rails/pull/42025/files#diff-6cbd9119d9ff9196372e53699cb973838036a376995cf32e4c0c5d6584520d6aR219-R237