@fxn 2021-04-22
Exactly. If we had versions, I'd expect a high-level code that would look like this:

```ruby
def load(payload)
  case version(payload)
when 6.1
  # Load with this format.
when 7.0
  # Load with this format.
else
  raise
end   
```

Then, each format branches compressed/uncompressed. That is just pseudocode to explain the mismatch between the naming and the encoding of the coder version the constants implicitly embed, not saying the actual implementation has to look like that.

If we encode two things in the same payload, I believe naming or branching should reflect that somehow.

Anyway, that is intuitive things I see on a superficial read, let me do the detailed one.