@fxn 2021-04-22
The exercise would be to imagine how would introducing 7.2 would look like.