@fxn 2021-04-22
For example:

```ruby
FORMAT_VERSION_6_0 = "\x04".b.freeze
FORMAT_VERSION_7_0 = "\x05".b.freeze
```

Now, in the case of 7.0, a second byte means compressed/uncompressed. In the case of 7.1, the 7.1 coder know where's the metadata and what it means.

That's the kind of things that come to mind.