@casperisfine 2021-04-22
I understand what you are saying, but I don't think it's worth splitting the two marker bytes. One byte gives us 255 possible versions, so we can put the compression bit in there as well.

Having both information in the same one byte prefix avoid having to slice the string twice.