@fxn 2021-04-22
Yes, you can make things in a way that use only one byte, and all parametrization of all version coders has to be globally unique. In that sense I said the implementation does not need to be that way. But, naming is not talking about versions, I read the code and see `MARSHAL_PREFIX` (indirection), not `FORMAT_VERSION_6_0` (meaning is obvious).

If you introduce 7.1, then I'd expect the `if/else` to read

```ruby
byte = payload.byteslice(0)
if byte == FORMAT_VERSION_6_1
  # ...
elsif byte == FORMAT_VERSION_7_0_COMPRESSED || byte == FORMAT_VERSION_7_0_UNCOMPRESSED
  # ...
else
  # ...
end
```

Again, example only to try to explain what I have in mind, that we are introducing coder versions, and the code does not have versions prominently in its names/logic so that it reads naturally with the mental model.