@fxn 2021-04-22
Name of coders and interface could also reflect that. So, instead of `LegacyCoder`, `Coder61` and `Coder70`. They provide a uniform interface `read/write`, they provide predicates `handles?(payload)`, etc.

I need to think about the need for a coder to be able to handle the payloads of other coders. That's a consequence of how things are designed, but the alternative in which a coder reads and writes its own format and a higher-level branch decides who handles a payload makes sense a priori too.

Need to work :(, will continue when I can.