@casperisfine 2021-04-22
@fxn I've refactored the PR quite a bit based on your recommendation, and also taken some parts out to hopefully make it easier to review.

Re:

> NB: The major change since #40412, is that except for the "compression byte", all the header is marshaled and compressed. This is mostly for simplicity's sake. If anything, splitting the metadata (expires_at, version) would allow to rebuild the Entry object without decompressing the payload, which in some rare scenarios could improve performance.

I'm starting to wonder splitting the value and the metadata could be much faster for string values, as we could skip `Marshal` entirely. If I have the time I'll try to benchmark that today or tomorrow.