@fxn 2021-04-22
@casperisfine fantastic man! Also, please feel free to leave guides docs for a followup PR, there's no point on maintaining docs for a patch still a bit in flux.