@casperisfine 2021-04-22
> I'm starting to wonder splitting the value and the metadata could be much faster for string values

So I tried it, and let's say that the results are mixed: https://gist.github.com/casperisfine/f6d2728ba91721929e775bc14af7438f, and that's before even trying to make sure the strings are returned with the proper encoding. So I'm not so sure it's really worth it.