@fxn 2021-04-24
@casperisfine Last refactor is great. I'd like to suggest only one final tweak.

These constants store format version marks ( or prefixes, signatures, magic bytes, ...):

```ruby
MARSHAL_PREFIX = "\x04\b".b.freeze
UNCOMPRESSED = "\x00".b.freeze
COMPRESSED = "\x01".b.freeze
```

They have to have distinct values, but the names are not distinct enough for my taste. For example, a marshaled payload is uncompressed too (overlap), and the relationship with the format version is not obvious right there, you need to _infer_ it from surrounding code.

In my view, those constants have to tell to the reader "I am a byte mark for this format version + variant". I believe names like the following would be more concise and make people think less:

```ruby
MARK_61              = "\x04\b".b.freeze # The one set by Marshal.
MARK_70_UNCOMPRESSED = "\x00".b.freeze
MARK_70_COMPRESSED   = "\x01".b.freeze
```

"MARK" could be something else if you prefer another choice.

WDYT?